pharma/backend/requirements.txt
fastapi
uvicorn 
pandas
python-dotenv
numpy

pyodbc